Move python related functions
Also the functions are slightly changed to return full path of the found
python executable.
